perf(get-selector): split by data locality into subsystem files
Break up the single 1,294-line file into five files under
`lib/core/utils/get-selector/` so each subsystem's data structures
live alongside the code that reads and writes them.

- `attr-format.js` (~105 lines): stateless attr formatters —
  `escapeAttribute`, `getAttributeNameValue`, `filterAttributes`,
  `ignoredAttributes`, `MAXATTRIBUTELENGTH`, control-char regex.
- `features.js` (~166 lines): read-only helpers — `getBaseSelector`,
  `getElmId`, `getThreeLeastCommonFeatures` (plus private
  `uncommonClasses`, `uncommonAttributes`, `countSort`).
- `bit-index.js` (~353 lines): every bit-work operation in one
  place — `listToBits`, `bitsForFragment`, `indicesForFragment`,
  `fragmentMatchesIdx`, `advanceAndFilter`. Treats `_perNode` as
  opaque read-only; owns `_matchCache` and the `Uint32Array`
  bitmap machinery.
- `precompute.js` (~375 lines): the one-time DOM walk —
  `getSelectorData`, `collectNodeAttrs`, `nthChildMap`. Owns
  `_perNode` and every field on it, plus every `data.*` structure
  the walk populates; calls `bit-index.listToBits` at finalize.
- `generate.js` (~303 lines): the per-target query phase —
  `computeFragment` + `generateSelector`. Owns `_computeFragmentCache`
  and `_chainCache` (populated at query time, not part of the
  precomputed model).

`lib/core/utils/get-selector.js` becomes a 20-line shim that
re-exports `getSelector` (memoized default) and `getSelectorData`,
so the import path `./get-selector` and the `axe.utils.getSelector`
/ `axe._selectorData` surfaces are unchanged. No other file in the
codebase needs to be touched.

Dependency graph is one-way: `precompute` and `generate` both
depend on `bit-index`; `generate` also depends on `features`;
`precompute` also depends on `attr-format` and `features`;
`get-selector.js` depends on `precompute` and `generate`. No
cycles.

Selector output remains byte-identical to develop across all 8
verification pages (mdn-live, stackoverflow, broken-workshop,
dream, webtestingcourse, mars, tensorflow, and the SEC 10-Q
filing: 2,480,657 rows compared). All 51 get-selector unit tests
still pass.
